Add files via upload
Create a Web application that displays the distance from NY to major cities in the world. The Web page allows the user to select one of the cities in the list shown below which is kept in a XML document. After selecting a city, then the app displays the corresponding distance.